refactor: split custom editors for VRC Constraint stubs and upgrade with UI versioning and shared style
