+++ VRCSDK3Stub/VRCPConStub/Components/VRCAimConstraint.cs
@@ -1,8 +1,4 @@
 using UnityEngine;
-using UnityEditor;
-using UnityEngine.UIElements;
-using UnityEditor.UIElements;
-using UnityEngine.Animations;
 using VRC.Dynamics.ManagedTypes;
 using System;
 using StubVersion = uk.novavoidhowl.dev.cvrfury.VRCPConStub.Common.StubVersion;
@@ -22,112 +18,6 @@ namespace VRC.SDK3.Dynamics.Constraint.Components
     public string StubVersion
     {
       get { return _stubVersion ?? uk.novavoidhowl.dev.cvrfury.VRCPConStub.Common.StubVersion.CurrentVersion; }
-    }
-  }
-
-  [CustomEditor(typeof(VRCAimConstraint))]
-  public class VRCAimConstraintEditor : Editor
-  {
-    public override VisualElement CreateInspectorGUI()
-    {
-      var root = new VisualElement();
-
-      var warningBox = new Box();
-      warningBox.style.marginTop = new StyleLength(10);
-      warningBox.style.paddingTop = new StyleLength(6);
-      warningBox.style.paddingBottom = new StyleLength(6);
-      warningBox.style.paddingLeft = new StyleLength(6);
-      warningBox.style.paddingRight = new StyleLength(6);
-      warningBox.style.backgroundColor = new StyleColor(new Color(1f, 0.8f, 0.8f, 0.3f));
-
-      var warningLabel = new Label(
-        "This VRChat constraint component should be converted to a standard Unity AimConstraint for compatibility with CVRFury."
-      );
-      warningLabel.style.whiteSpace = WhiteSpace.Normal;
-      warningBox.Add(warningLabel);
-
-      var convertButton = new Button(() =>
-      {
-        ConvertToUnityConstraint();
-      })
-      {
-        text = "Convert to Unity AimConstraint"
-      };
-      convertButton.style.marginTop = new StyleLength(10);
-
-      root.Add(warningBox);
-      root.Add(convertButton);
-
-      return root;
-    }
-
-    private void ConvertToUnityConstraint()
-    {
-      VRCAimConstraint vrcConstraint = (VRCAimConstraint)target;
-      GameObject gameObject = vrcConstraint.gameObject;
-
-      // Add Unity AimConstraint component
-      AimConstraint unityConstraint = Undo.AddComponent<AimConstraint>(gameObject);
-
-      // Transfer basic properties
-      unityConstraint.weight = vrcConstraint.GlobalWeight;
-      unityConstraint.constraintActive = vrcConstraint.IsActive;
-
-      // Set aim vector
-      unityConstraint.aimVector = vrcConstraint.AimAxis;
-      unityConstraint.upVector = vrcConstraint.UpAxis;
-
-      // Set world up type
-      switch (vrcConstraint.WorldUp)
-      {
-        case VRC.Dynamics.ManagedTypes.VRCConstraintBase.WorldUpType.SceneUp:
-          unityConstraint.worldUpType = AimConstraint.WorldUpType.SceneUp;
-          break;
-        case VRC.Dynamics.ManagedTypes.VRCConstraintBase.WorldUpType.ObjectUp:
-          unityConstraint.worldUpType = AimConstraint.WorldUpType.ObjectUp;
-          unityConstraint.worldUpObject = vrcConstraint.WorldUpTransform;
-          break;
-        case VRC.Dynamics.ManagedTypes.VRCConstraintBase.WorldUpType.ObjectRotationUp:
-          unityConstraint.worldUpType = AimConstraint.WorldUpType.ObjectRotationUp;
-          unityConstraint.worldUpObject = vrcConstraint.WorldUpTransform;
-          break;
-        case VRC.Dynamics.ManagedTypes.VRCConstraintBase.WorldUpType.Vector:
-          unityConstraint.worldUpType = AimConstraint.WorldUpType.Vector;
-          unityConstraint.worldUpVector = vrcConstraint.WorldUpVector;
-          break;
-        case VRC.Dynamics.ManagedTypes.VRCConstraintBase.WorldUpType.None:
-          unityConstraint.worldUpType = AimConstraint.WorldUpType.None;
-          break;
-      }
-
-      // Add sources
-      for (int i = 0; i < vrcConstraint.Sources.Count; i++)
-      {
-        var vrcSource = vrcConstraint.Sources[i];
-        if (vrcSource.SourceTransform != null)
-        {
-          ConstraintSource unitySource = new ConstraintSource
-          {
-            sourceTransform = vrcSource.SourceTransform,
-            weight = vrcSource.Weight
-          };
-          unityConstraint.AddSource(unitySource);
-        }
-      }
-
-      // Set axes
-      unityConstraint.rotationAxis = 0;
-      if (vrcConstraint.AffectsRotationX)
-        unityConstraint.rotationAxis |= Axis.X;
-      if (vrcConstraint.AffectsRotationY)
-        unityConstraint.rotationAxis |= Axis.Y;
-      if (vrcConstraint.AffectsRotationZ)
-        unityConstraint.rotationAxis |= Axis.Z;
-
-      // Remove VRC component after conversion
-      Undo.DestroyObjectImmediate(vrcConstraint);
-
-      EditorUtility.DisplayDialog("Conversion Complete", "Successfully converted to Unity AimConstraint", "OK");
     }
   }
 }
