+++ VRCSDK3Stub/VRCPConStub/Components/VRCParentConstraint.cs
@@ -1,8 +1,4 @@
 using UnityEngine;
-using UnityEditor;
-using UnityEngine.UIElements;
-using UnityEditor.UIElements;
-using UnityEngine.Animations;
 using VRC.Dynamics.ManagedTypes;
 using System;
 using StubVersion = uk.novavoidhowl.dev.cvrfury.VRCPConStub.Common.StubVersion;
@@ -22,103 +18,6 @@ namespace VRC.SDK3.Dynamics.Constraint.Components
     public string StubVersion
     {
       get { return _stubVersion ?? uk.novavoidhowl.dev.cvrfury.VRCPConStub.Common.StubVersion.CurrentVersion; }
-    }
-  }
-
-  [CustomEditor(typeof(VRCParentConstraint))]
-  public class VRCParentConstraintEditor : Editor
-  {
-    public override VisualElement CreateInspectorGUI()
-    {
-      var root = new VisualElement();
-
-      var warningBox = new Box();
-      warningBox.style.marginTop = new StyleLength(10);
-      warningBox.style.paddingTop = new StyleLength(6);
-      warningBox.style.paddingBottom = new StyleLength(6);
-      warningBox.style.paddingLeft = new StyleLength(6);
-      warningBox.style.paddingRight = new StyleLength(6);
-      warningBox.style.backgroundColor = new StyleColor(new Color(1f, 0.8f, 0.8f, 0.3f));
-
-      var warningLabel = new Label(
-        "This VRChat constraint component should be converted to a standard Unity ParentConstraint for compatibility with CVRFury."
-      );
-      warningLabel.style.whiteSpace = WhiteSpace.Normal;
-      warningBox.Add(warningLabel);
-
-      var convertButton = new Button(() =>
-      {
-        ConvertToUnityConstraint();
-      })
-      {
-        text = "Convert to Unity ParentConstraint"
-      };
-      convertButton.style.marginTop = new StyleLength(10);
-
-      root.Add(warningBox);
-      root.Add(convertButton);
-
-      return root;
-    }
-
-    private void ConvertToUnityConstraint()
-    {
-      VRCParentConstraint vrcConstraint = (VRCParentConstraint)target;
-      GameObject gameObject = vrcConstraint.gameObject;
-
-      // Add Unity ParentConstraint component
-      ParentConstraint unityConstraint = Undo.AddComponent<ParentConstraint>(gameObject);
-
-      // Transfer basic properties
-      unityConstraint.weight = vrcConstraint.GlobalWeight;
-      unityConstraint.constraintActive = vrcConstraint.IsActive;
-
-      // Set translation offset
-      unityConstraint.translationAtRest = vrcConstraint.PositionAtRest;
-
-      // Set rotation offset
-      unityConstraint.rotationAtRest = vrcConstraint.RotationAtRest;
-
-      // Add sources
-      for (int i = 0; i < vrcConstraint.Sources.Count; i++)
-      {
-        var vrcSource = vrcConstraint.Sources[i];
-        if (vrcSource.SourceTransform != null)
-        {
-          ConstraintSource unitySource = new ConstraintSource
-          {
-            sourceTransform = vrcSource.SourceTransform,
-            weight = vrcSource.Weight
-          };
-          int sourceIndex = unityConstraint.AddSource(unitySource);
-
-          // Set position and rotation offsets for this source
-          unityConstraint.SetTranslationOffset(sourceIndex, vrcSource.ParentPositionOffset);
-          unityConstraint.SetRotationOffset(sourceIndex, vrcSource.ParentRotationOffset);
-        }
-      }
-
-      // Set axes
-      unityConstraint.translationAxis = 0;
-      if (vrcConstraint.AffectsPositionX)
-        unityConstraint.translationAxis |= Axis.X;
-      if (vrcConstraint.AffectsPositionY)
-        unityConstraint.translationAxis |= Axis.Y;
-      if (vrcConstraint.AffectsPositionZ)
-        unityConstraint.translationAxis |= Axis.Z;
-
-      unityConstraint.rotationAxis = 0;
-      if (vrcConstraint.AffectsRotationX)
-        unityConstraint.rotationAxis |= Axis.X;
-      if (vrcConstraint.AffectsRotationY)
-        unityConstraint.rotationAxis |= Axis.Y;
-      if (vrcConstraint.AffectsRotationZ)
-        unityConstraint.rotationAxis |= Axis.Z;
-
-      // Remove VRC component after conversion
-      Undo.DestroyObjectImmediate(vrcConstraint);
-
-      EditorUtility.DisplayDialog("Conversion Complete", "Successfully converted to Unity ParentConstraint", "OK");
     }
   }
 }
