+++ VRCStubConvertAndUI/VRCPConUIAndConverter/Common/UIVersion.cs
@@ -1,0 +1,26 @@
+using System;
+
+namespace VRC.SDK3.Dynamics.Constraint.Editors.Common
+{
+  public static class UIVersion
+  {
+    private static readonly Version _version = new Version(2, 0, 0);
+
+    // For serialization and display
+    public static string CurrentVersion => _version.ToString();
+
+    // For version comparison
+    public static Version AsVersion => _version;
+
+    // Helper method for version checking
+    public static bool IsVersionAtLeast(Version otherVersion)
+    {
+      return _version.CompareTo(otherVersion) >= 0;
+    }
+
+    public static bool IsVersionAtLeast(string otherVersion)
+    {
+      return IsVersionAtLeast(new Version(otherVersion));
+    }
+  }
+}
