+++ VRCStubConvertAndUI/VRCPConUIAndConverter/VRCPConUIAndConverter.csproj
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{A1B2C3D4-E5F6-7890-ABCD-1234567890A6}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>VRC.SDK3.Dynamics.Constraint.Editors</RootNamespace>
+    <AssemblyName>VRCPConConverter</AssemblyName>
+    <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
+    <FileAlignment>512</FileAlignment>
+    <Deterministic>true</Deterministic>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>..\Build\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
+    <Reference Include="UnityEngine">
+      <HintPath>C:\Program Files\Unity\Hub\Editor\2021.3.45f1\Editor\Data\Managed\UnityEngine.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="UnityEditor">
+      <HintPath>C:\Program Files\Unity\Hub\Editor\2021.3.45f1\Editor\Data\Managed\UnityEditor.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <!-- Reference the corresponding stub DLL -->
+    <Reference Include="VRCPConStub">
+      <HintPath>..\..\VRCSDK3Stub\Build\VRCPConStub.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="VRCPstub">
+      <HintPath>..\..\VRCSDK3Stub\Build\VRCPstub.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <!-- Reference shared UI styles -->
+    <Reference Include="VRCStubUICore">
+      <HintPath>..\Build\VRCStubUICore.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Common\UIVersion.cs" />
+    <Compile Include="Editors\VRCAimConstraintEditor.cs" />
+    <Compile Include="Editors\VRCLookAtConstraintEditor.cs" />
+    <Compile Include="Editors\VRCParentConstraintEditor.cs" />
+    <Compile Include="Editors\VRCPositionConstraintEditor.cs" />
+    <Compile Include="Editors\VRCRotationConstraintEditor.cs" />
+    <Compile Include="Editors\VRCScaleConstraintEditor.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Common\" />
+    <Folder Include="Editors\" />
+    <Folder Include="Resources\" />
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <Target Name="PostBuildEvent">
+    <Exec Command="echo Built VRCPConConverter.dll - VRC Constraint Editors and Converters" />
+  </Target>
+</Project>
